updated main screen(login screen) to match the newest version

+++ app/src/main/java/com/example/mbank/utils/UIUtils.java
@@ -46,11 +46,11 @@ public class UIUtils {
         }
     }
 
-    public static int dpToPx(int dp, Context context) {
+    public static int dpToPx(float dp, Context context) {
         float density = context.getResources()
                 .getDisplayMetrics()
                 .density;
-        return Math.round((float) dp * density);
+        return Math.round(dp * density);
     }
 
     public static int getStatusBarHeight(Context context) {
